User models Update

+++ src/app/beer_garden/api/http/handlers/vbeta/file.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 import io
 
-from brewtils.models import Resolvable
+from brewtils.models import Permissions, Resolvable
 from brewtils.schema_parser import SchemaParser
 
-from beer_garden.api.http.base_handler import BaseHandler
+from beer_garden.api.http.handlers import AuthorizationHandler
 from beer_garden.db.mongo.models import RawFile
+from beer_garden.garden import local_garden
 
 
-class RawFileAPI(BaseHandler):
+class RawFileAPI(AuthorizationHandler):
     async def get(self, file_id):
         """
         ---
@@ -31,6 +32,8 @@ class RawFileAPI(BaseHandler):
         tags:
           - Files
         """
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile.objects.get(id=file_id)
         file = db_file.file.read()
 
@@ -59,6 +62,9 @@ class RawFileAPI(BaseHandler):
         tags:
           - Files
         """
+        self.minimum_permission = Permissions.OPERATOR.name
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile.objects.get(id=file_id)
         db_file.file.delete()
         db_file.save()
@@ -66,7 +72,7 @@ class RawFileAPI(BaseHandler):
         self.set_status(204)
 
 
-class RawFileListAPI(BaseHandler):
+class RawFileListAPI(AuthorizationHandler):
     async def post(self):
         """
         ---
@@ -88,6 +94,9 @@ class RawFileListAPI(BaseHandler):
         tags:
           - Files
         """
+        self.minimum_permission = Permissions.OPERATOR.name
+
+        self.verify_user_permission_for_object(local_garden())
         db_file = RawFile()
         db_file.file.put(io.BytesIO(self.request.body))
         db_file.save()
